Fix for Task 231
