+++ plugins/org.eclipse.osee.orcs.rest/src/org/eclipse/osee/orcs/rest/internal/SavedSearchEndpoint.java
@@ -31,7 +31,6 @@ import org.eclipse.osee.framework.core.enums.CoreAttributeTypes;
 import org.eclipse.osee.framework.core.enums.CoreBranches;
 import org.eclipse.osee.orcs.OrcsApi;
 import org.eclipse.osee.orcs.rest.model.search.artifact.SavedSearch;
-import org.eclipse.osee.orcs.rest.model.search.artifact.SavedSearchRequest;
 import org.eclipse.osee.orcs.transaction.TransactionBuilder;
 
 @Path("/savedSearch")
@@ -47,18 +46,19 @@ public class SavedSearchEndpoint {
    @POST
    @Consumes(MediaType.APPLICATION_JSON)
    @Produces(MediaType.APPLICATION_JSON)
-   public Response createSavedSearch(SavedSearchRequest savedSearchRequest) {
-      validateSavedSearch(savedSearchRequest);
+   public Response createSavedSearch(SavedSearch savedSearch) {
+      validateSavedSearch(savedSearch);
       UserId currentUser = getCurrentUserId();
 
-      String payload = toPayload(savedSearchRequest);
+      savedSearch.setId(null);
+      String payload = toPayload(savedSearch);
 
       try {
          TransactionBuilder tx =
             orcsApi.getTransactionFactory().createTransaction(CoreBranches.COMMON, "Create Saved Search");
          AttributeId attributeId = tx.createAttribute(currentUser, CoreAttributeTypes.SavedSearch, payload);
          tx.commit();
-         SavedSearch savedSearch = toSavedSearch(savedSearchRequest, attributeId.getId());
+         savedSearch.setId(attributeId.getId());
          return Response.ok(savedSearch).build();
       } catch (Exception ex) {
          throw new WebApplicationException("Error creating SavedSearch", ex, Status.INTERNAL_SERVER_ERROR);
@@ -137,12 +137,12 @@ public class SavedSearchEndpoint {
    @Path("/{id}")
    @Consumes(MediaType.APPLICATION_JSON)
    @Produces(MediaType.APPLICATION_JSON)
-   public Response updateSavedSearch(@PathParam("id") Long id, SavedSearchRequest savedSearchRequest) {
+   public Response updateSavedSearch(@PathParam("id") Long id, SavedSearch savedSearch) {
       if (id == null || id <= 0) {
          throw new WebApplicationException("id is required", Status.BAD_REQUEST);
       }
 
-      validateSavedSearch(savedSearchRequest);
+      validateSavedSearch(savedSearch);
       UserId currentUser = getCurrentUserId();
 
       try {
@@ -165,13 +165,13 @@ public class SavedSearchEndpoint {
             throw new WebApplicationException("SavedSearch not found", Status.NOT_FOUND);
          }
 
-         String payload = toPayload(savedSearchRequest);
+         String payload = toPayload(savedSearch);
          TransactionBuilder tx =
             orcsApi.getTransactionFactory().createTransaction(CoreBranches.COMMON, "Update Saved Search");
          tx.setAttributeById(currentUser, AttributeId.valueOf(id), payload);
          tx.commit();
 
-         SavedSearch savedSearch = toSavedSearch(savedSearchRequest, id);
+         savedSearch.setId(id);
          return Response.ok(savedSearch).build();
       } catch (WebApplicationException wae) {
          throw wae;
@@ -180,29 +180,29 @@ public class SavedSearchEndpoint {
       }
    }
 
-   private void validateSavedSearch(SavedSearchRequest savedSearchRequest) {
-      if (savedSearchRequest == null || savedSearchRequest.getTitle() == null || savedSearchRequest.getTitle().trim().isEmpty()) {
+   private void validateSavedSearch(SavedSearch savedSearch) {
+      if (savedSearch == null || savedSearch.getTitle() == null || savedSearch.getTitle().trim().isEmpty()) {
          throw new WebApplicationException("title is required", Status.BAD_REQUEST);
       }
-      if (savedSearchRequest.getQuery() == null || savedSearchRequest.getQuery().trim().isEmpty()) {
+      if (savedSearch.getQuery() == null || savedSearch.getQuery().trim().isEmpty()) {
          throw new WebApplicationException("query is required", Status.BAD_REQUEST);
       }
-      savedSearchRequest.setTitle(savedSearchRequest.getTitle().trim());
-      savedSearchRequest.setQuery(savedSearchRequest.getQuery().trim());
-      if (savedSearchRequest.getTimestamp() == null) {
-         savedSearchRequest.setTimestamp(System.currentTimeMillis());
+      savedSearch.setTitle(savedSearch.getTitle().trim());
+      savedSearch.setQuery(savedSearch.getQuery().trim());
+      if (savedSearch.getTimestamp() == null) {
+         savedSearch.setTimestamp(System.currentTimeMillis());
       }
-      if (savedSearchRequest.getArtifactTypes() == null) {
-         savedSearchRequest.setArtifactTypes(new ArrayList<>());
+      if (savedSearch.getArtifactTypes() == null) {
+         savedSearch.setArtifactTypes(new ArrayList<>());
       }
-      if (savedSearchRequest.getAttributeTypes() == null) {
-         savedSearchRequest.setAttributeTypes(new ArrayList<>());
+      if (savedSearch.getAttributeTypes() == null) {
+         savedSearch.setAttributeTypes(new ArrayList<>());
       }
-      if (savedSearchRequest.getExactMatch() == null) {
-         savedSearchRequest.setExactMatch(Boolean.FALSE);
+      if (savedSearch.getExactMatch() == null) {
+         savedSearch.setExactMatch(Boolean.FALSE);
       }
-      if (savedSearchRequest.getSearchById() == null) {
-         savedSearchRequest.setSearchById(Boolean.FALSE);
+      if (savedSearch.getSearchById() == null) {
+         savedSearch.setSearchById(Boolean.FALSE);
       }
    }
 
@@ -214,19 +214,12 @@ public class SavedSearchEndpoint {
       return currentUser;
    }
 
-   private String toPayload(SavedSearchRequest savedSearchRequest) {
+   private String toPayload(SavedSearch savedSearch) {
       try {
-         return mapper.writeValueAsString(new SavedSearchPayload(savedSearchRequest));
+         return mapper.writeValueAsString(new SavedSearchPayload(savedSearch));
       } catch (Exception e) {
          throw new WebApplicationException("savedSearch serialization failed", e, Status.BAD_REQUEST);
       }
-   }
-
-   private SavedSearch toSavedSearch(SavedSearchRequest savedSearchRequest, Long id) {
-      SavedSearch savedSearch =
-         new SavedSearch(savedSearchRequest.getTitle(), savedSearchRequest.getQuery(), savedSearchRequest.getTimestamp());
-      savedSearch.setId(id);
-      return savedSearch;
    }
 
    private SavedSearch fromPayload(String payload, Long attributeId) {
@@ -243,19 +236,19 @@ public class SavedSearchEndpoint {
       public final String title;
       public final String query;
       public final Long timestamp;
-      public final List<SavedSearchRequest.SavedSearchSelection> artifactTypes;
-      public final List<SavedSearchRequest.SavedSearchSelection> attributeTypes;
+      public final List<SavedSearch.SavedSearchSelection> artifactTypes;
+      public final List<SavedSearch.SavedSearchSelection> attributeTypes;
       public final Boolean exactMatch;
       public final Boolean searchById;
 
-      private SavedSearchPayload(SavedSearchRequest savedSearchRequest) {
-         this.title = savedSearchRequest.getTitle();
-         this.query = savedSearchRequest.getQuery();
-         this.timestamp = savedSearchRequest.getTimestamp();
-         this.artifactTypes = savedSearchRequest.getArtifactTypes();
-         this.attributeTypes = savedSearchRequest.getAttributeTypes();
-         this.exactMatch = savedSearchRequest.getExactMatch();
-         this.searchById = savedSearchRequest.getSearchById();
+      private SavedSearchPayload(SavedSearch savedSearch) {
+         this.title = savedSearch.getTitle();
+         this.query = savedSearch.getQuery();
+         this.timestamp = savedSearch.getTimestamp();
+         this.artifactTypes = savedSearch.getArtifactTypes();
+         this.attributeTypes = savedSearch.getAttributeTypes();
+         this.exactMatch = savedSearch.getExactMatch();
+         this.searchById = savedSearch.getSearchById();
       }
    }
 }
