+++ web/apps/osee/src/app/ple/artifact-explorer/advanced-search-page/advanced-search-page.component.ts
@@ -143,9 +143,6 @@ type SavedSearch = {
 	title: string;
 	query: string;
 	timestamp?: number;
-};
-
-type SavedSearchRequest = SavedSearch & {
 	artifactTypes?: NamedId[];
 	attributeTypes?: NamedId[];
 	exactMatch?: boolean;
@@ -1767,7 +1764,7 @@ export class AdvancedSearchPageComponent implements OnInit {
 		}
 		this.editErrorMessage = '';
 		this.editSaveInProgress = true;
-		const updatedSearch: SavedSearchRequest = {
+		const updatedSearch: SavedSearch = {
 			...savedSearch,
 			title: updatedTitle,
 			query: (this.editingSearchQuery || '').trim(),
